+++ CHANGELOG.rst
@@ -1,3 +1,8 @@
+next
+====
+
+- Support pdf MaxVersion up to 1.7 (if the underlying cairo supports it).
+
 v0.5 (2022-08-18)
 =================